modified screen names and separated transcript from parent
